+++ src/rack/middleware/cache.js
@@ -11,51 +11,42 @@ export class CacheMiddleware extends KinveyMiddleware {
     this.adapters = adapters;
   }
 
-  handle(request) {
-    return super.handle(request).then(() => {
-      const method = request.method;
-      const query = request.query;
-      const data = request.data;
-      const db = new DB(request.appKey, this.adapters);
-      let promise;
+  async handle(request) {
+    request = await super.handle(request);
+    const method = request.method;
+    const query = request.query;
+    const data = request.data;
+    const db = new DB(request.appKey, this.adapters);
+    let result;
 
-      if (method === HttpMethod.GET) {
-        if (request.entityId) {
-          if (request.entityId === '_count') {
-            promise = db.count(request.collectionName, query);
-          } else if (request.entityId === '_group') {
-            promise = db.group(request.collectionName, data);
-          } else {
-            promise = db.findById(request.collectionName, request.entityId);
-          }
+    if (method === HttpMethod.GET) {
+      if (request.entityId) {
+        if (request.entityId === '_count') {
+          result = await db.count(request.collectionName, query);
+        } else if (request.entityId === '_group') {
+          result = await db.group(request.collectionName, data);
         } else {
-          promise = db.find(request.collectionName, query);
+          result = await db.findById(request.collectionName, request.entityId);
         }
-      } else if (method === HttpMethod.POST || method === HttpMethod.PUT) {
-        promise = db.save(request.collectionName, data);
-      } else if (method === HttpMethod.DELETE) {
-        if (request.entityId) {
-          promise = db.removeById(request.collectionName, request.entityId);
-        } else {
-          promise = db.remove(request.collectionName, query);
-        }
+      } else {
+        result = await db.find(request.collectionName, query);
       }
+    } else if (method === HttpMethod.POST || method === HttpMethod.PUT) {
+      result = await db.save(request.collectionName, data);
+    } else if (method === HttpMethod.DELETE) {
+      if (request.entityId) {
+        result = await db.removeById(request.collectionName, request.entityId);
+      } else {
+        result = await db.remove(request.collectionName, query);
+      }
+    }
 
-      return promise.then(result => {
-        let statusCode = StatusCode.Ok;
+    request.response = {
+      statusCode: method === HttpMethod.POST ? StatusCode.Created : StatusCode.Ok,
+      headers: {},
+      data: result
+    };
 
-        if (method === HttpMethod.POST) {
-          statusCode = StatusCode.Created;
-        }
-
-        request.response = {
-          statusCode: statusCode,
-          headers: {},
-          data: result
-        };
-
-        return request;
-      });
-    });
+    return request;
   }
 }