+++ src/rack/middleware/parse.js
@@ -8,25 +8,24 @@ export class ParseMiddleware extends KinveyMiddleware {
     super(name);
   }
 
-  handle(request) {
-    return super.handle(request).then(() => {
-      const response = request.response;
+  async handle(request) {
+    request = await super.handle(request);
+    const response = request.response;
 
-      if (response && response.data) {
-        const contentType = response.headers['content-type'] || response.headers['Content-Type'];
+    if (response && response.data) {
+      const contentType = response.headers['content-type'] || response.headers['Content-Type'];
 
-        if (contentType.indexOf('application/json') === 0) {
-          try {
-            response.data = JSON.parse(response.data);
-          } catch (err) {
-            response.data = response.data;
-          }
-
-          request.response = response;
+      if (contentType.indexOf('application/json') === 0) {
+        try {
+          response.data = JSON.parse(response.data);
+        } catch (err) {
+          response.data = response.data;
         }
-      }
 
-      return request;
-    });
+        request.response = response;
+      }
+    }
+
+    return request;
   }
 }