+++ test/setup.js
@@ -3,27 +3,18 @@ import { randomString } from '../src/utils/string';
 import { NetworkRack } from '../src/rack/rack';
 import { SerializeMiddleware } from '../src/rack/middleware/serialize';
 import { HttpMiddleware } from './mocks/http';
-import { DeviceAdapter } from './mocks/device';
-import { Device } from '../src/utils/device';
 import sinon from 'sinon';
-
-before(function() {
-  Device.use(new DeviceAdapter());
-});
 
 before(function() {
   const networkRack = NetworkRack.sharedInstance();
   networkRack.useAfter(SerializeMiddleware, new HttpMiddleware());
 });
 
-
-beforeEach(function() {
-  this.sandbox = sinon.sandbox.create();
-});
-
-afterEach(function() {
-  this.sandbox.restore();
-  delete this.sandbox;
+before(function() {
+  this.client = Kinvey.init({
+    appKey: randomString(),
+    appSecret: randomString()
+  });
 });
 
 beforeEach(function() {
@@ -35,4 +26,13 @@ beforeEach(function() {
 
 afterEach(function() {
   delete this.client;
+});
+
+beforeEach(function() {
+  this.sandbox = sinon.sandbox.create();
+});
+
+afterEach(function() {
+  this.sandbox.restore();
+  delete this.sandbox;
 });